[Feat] 날짜와 시간을 표시하는 ListViewCell 구현
sampleData에서 추가했던 시간을 Cell에서 확인하도록 구현
